requirements.txt
uvicorn
fastapi
pydantic
tortoise-orm
aerich
aiomysql
python-dotenv